requirements.txt
exrex
numpy
transformers
tensorflow
unidecode
scikit-learn
pandas
wandb
python-dotenv
torch
accelerate